% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env2list.R
\name{env_matches}
\alias{env_matches}
\alias{env_starts_with}
\alias{env_ends_with}
\alias{env_contains}
\alias{env_num_range}
\title{Select objects from the global environment that match a pattern}
\usage{
env_matches(match, silent = FALSE, ...)

env_starts_with(match, silent = FALSE, ...)

env_ends_with(match, silent = FALSE, ...)

env_contains(match, silent = FALSE, ...)

env_num_range(prefix, range, silent = FALSE, ...)
}
\arguments{
\item{match}{A character vector.}

\item{silent}{Whether to message assigned objects in console.}

\item{...}{Additional parameters passed to `grep()`}

\item{prefix}{A prefix that starts the numeric range.}

\item{range}{A sequence of integers, like 1:5.}
}
\value{
a list
}
\description{
* [env_starts_with()]: Matches prefix.

* [env_ends_with()]: Matches suffix.

* [env_contains()]: Contains a literal string.

* [env_matches()]: Matches a regular expression.

* [env_num_range()]: Matches a numerical range like x01, x02, x03.
}
\examples{
\dontrun{
library(nycflights13)
data("airlines")
data("airports")
data("airmiles")
data("airquality")
air_dfs <- env_matches("^air")
names(air_dfs)
str(air_dfs, max.level = 1)
}
}
